Bug 693028 - improve mupdf viewer handling of broken page streams
Currently, if a page stream cannot be read, mupdf gives an alert box
and then exits. This is annoying when reading a large pdf.

Here we change the code to only exit if a page is completely broken;
in the case of missing page contents, or missing links, we give a warning
and just render the best we can.

Also, update a couple of error messages to be less misleading.