Fix bug in ARM code in draw_simple_scale
The scale_row_from_temp code was broken. Firstly the rounding was wrong
in the 'bulk' case (not a big deal), but more importantly on
configurations where unaligned loads were not allowed (such as the nook),
we could still crash due to an incorrect test to avoid that code.

Thanks to Kammerer for the report, and testing of fixed version.