mupdfclean - update to allow renumbering of encrypted objects
mupdfclean (or more correctly, the pdf_write function) currently has
a limitation, in that we cannot renumber objects when encryption is
being used. This is because the object/generation number is pickled
into the stream, and renumbering the object causes it to become
unreadable.

The solution used here is to provide extended functions that take both
the object/generation number and the original object/generation
number. The original object numbers are only used for setting up the
encryption.

pdf_write now keeps track of the original object/generation number
for each object.

This fix is important, if we ever want to output linearized pdf as
this requires us to be able to renumber objects to a very specific
order.

We also make a fix in removeduplicateobjects that should only
matter in the case where we fail to read an object correctly.