Fix Bug 693097; memory overwrite when handling knockout groups.
When handling knockout groups, we have to copy the background from the
previous group in so we can 'knockout' properly. If the previous group
is a different colorspace, this gives us problems!

The fix, implemented here, is to update the copy_pixmap_rect function
to know how to copy between pixmaps of different depth.

Gray <-> RGB are the ones we really care about; the generic code will
probably do a horrible job, but shouldn't ever be called at present.

This suffices to stop the crashing - we will probably revisit this
when we revise the blending support.