Add better mechanism for enumerating annotation rectangles.
Rather than having a dedicated call to enumerate the rectangles for the
annotations on a page, add an interface for enumerating annotations
with accessor functions. Currently the only accessor function is
the one to get the annotation rectangle.

Use this new scheme in place of fz_bound_annots within mudraw.

Also use this scheme to set the caret cursor in the viewer when over
a data field.