Updates for more effective mujstest.
Mudraw is updated so when the -j flag is used it guesses the length
of the text it should output for each field. It also outputs a
"Lorem Ipsum" string, which should hopefully test the layout
facilities a bit better.

Mujstest is updated so it can read the escaped chars from that
string.