Bug 693160: Fix bug in fz_write_buffer_bits.
When writing few enough bits that they would fit into the 'spare' bits
in the last byte, I was failing to update the buffer.

Also, I was failing to grow the buffer enough, and calculating the
wrong number of bits left over in various places.

Both fixed here. Thanks to Robert Jedrzejczyk and Sebras!