Improve detection of images for pdfwrite.
In pdfwrite we have a flag to say "don't expand images", but this
isn't always honoured because we don't know that a stream is an
image (for instance the corrupt fax stream used as a thumbnail
in normal_439.pdf).

Here we update the code to spot that streams are likely to be
images based on the filters in use, or on the presence of both
Width and Height tags.