Fix comparison of cmap table length against maximum value
The cmap table length counts how many entries there are currently in
the table. The table length was previously tested against USHRT_MAX
which is 65535. However, the desired value to compare with was 65536
which would be the maximum number of entries allowed in a cmap table.
All comparisons of the cmap table length are now using USHRT_MAX + 1.