Fix various problems with linearisation when saving.
Unused objects could cause problems with the sort order and picking
the object to start with. Now coped with.

If the hintstream object replaces another object that already had a
stream, pdf_open_raw_filter would get confused by the presence of a
stm_buf. Now fixed.

Fix a 64bit problem in page_objects_list_ensure, as well as tweaking
the code for readability.

When outputting single page files, we can end up with opts->start = 1
and this upset the offset calculating logic.

Insist on compacting the xref when linearising.

Thanks to Sebras and Zeniko for providing test cases. This commit
should (hopefully) stop the SEGVs, but there are still cases where
Acrobat doesn't think that the files output are "Optimised for Fast
Web View". I cannot see why.