Fix bug 693229: Mupdf/android has inconsistent search phrase highlighting
The highlights were unintentionally being cached between file invocations.
In fact it was possible for the highlighting from one file to appear when
opening another, even if those highlights didn't match words on the page.
That could happen if both files happened to have been last opened on a
common page (common in terms of page number).