Handle missing clear codes in LZW.
Previously if the lookup table was full and no clear code appeared the
decoder would try to add more entries to the table. However the table
is of fixed size (4096 entries) so it would write outside the table.

Fix this by detecting when the lookup table is full and a clear code
ought to appear. At this point the decoder will now treat and process
any code as a clear code.

For valid documents this will never happen, for invalid documents this
means risking that succeeding codes may be misinterpreted and that the
decoded data will be incorrect, this case should be handled by the
consumer of the data though.

Fixes bug 693306.