Forms: avoid javascript action execution when engine not available
This was necessary to avoid indirecting through a NULL pointer returned
from pdf_js_get_event, but is a generally sensible restriction.

Also separate the execution of the document-level javascript actions
from the pdf_js contstructor, so that doc->js is set during those
actions.

Also add a missing const