Bug 693350: Add some dreaded 'const's for string keys.
On the whole we avoid using const within MuPDF, but bug 693350
highlights cases where this can cause a problem with C++.

In C, if you do:  foo("bar");  then "bar" has type char *.

In C++, if you do foo("bar");  then "bar" has type const char *.

This means that any calls to the MuPDF library from C++ that take
strings give warnings.

The fix is simple, so it seems to be worthwhile adding a few consts.
None of our internal data structures are affected in any way by
this change.

Thanks to Franz Fellner for pointing out this issue.