Add fz_update_page
Regenerate dirty appearance streams and report changed annotations since
last call.

Also include a partial revert of changes in 96f335bc, that turn out not
to be necessary.

fz_update_page must now be called between each document-changing event and
the next render. pdfapp.c and the android app have been updated to do so,
but do not yet take advantage of the possibility to render only the updated
areas of the screen.