Android: fix rendering-wrong-page bug
The page number was being cached at two levels, and some calls
were skipping the upper level, leaving it out of date. The
fix is always to call the lower level gotoPage from the upper level.
Since the lower level caches, there is little to gain in avoiding
the call, in any case.