Update scan converter to cope better with rectangular paths.
Currently the scan converter advances one 'subpixel' scanline at a
time; here we update it to work in multiple subpixel scanlines at a
time.

If we spot that the gel consists of entirely vertical edges, then we
calculate the height for which those edges will remain unchanged. This
allows us to deal quickly with rectangular paths.

In the case of large vertical only edges, we can process multiple
scanlines (not just subpixel scanlines) at once.