Tweak image color conversion
Use just 1 loop rather than 2, and count downwards as this is faster
on most architectures.

For the 'hash tabled memoized' general case, the time taken to form the
hashes is significant. Add some code to check that the pixel isn't the
same as the one we just did and bypass the hash.