Fix fz_try/fz_catch in overflow case.
Thanks to zeniko for pointing out that the recent changes to
the fz_try/fz_catch macros to allow for throws in the fz_always
block had broken the exception stack overflow case.

Thanks also for the example file (nesting stack overflow.pdf),
which has now been added to the regression suite.