Bug 693290: Fix use after free in obj stream handling.
Thanks to zeniko for pointing this out. If we encounter a new definition
for a given object (presumably due to a repair operation), we used to
throw the old one away, and keep the new one. This could cause any
current holders of the object to be left with a stale pointer.

Now we throw the new one away and keep the old one - with a warning
if they are different.