Improve mutool clean behaviour on broken streams.
When cleaning a file with a corrupt stream in it, historically mupdf
would give up when it encountered such a stream. This is often not
what is desired, as information can be lost.

The changes herein allow us to use our best efforts when reading
a stream, so that broken streams are reproduced in the output
cleaned file.

Problem found in a test file, pdf_001/2599.pdf.asan.58.1778 supplied
by Mateusz "j00ru" Jurczyk and Gynvael Coldwind of the Google
Security Team using Address Sanitizer. Many thanks!