More work on android bitmap recycling.
After much discussion and investigation, Paul and I have realised
that we do in fact (in the current scheme at least) need to hold
the existing bitmap in memory while drawing the next one (as the
existing bitmap is still in an ImageView and being used for any
foreground render requests). As such remove a 'setBm(null)' in
drawPage.

Also, in the onPostExecute for the patch redraw, we cannot recycle
the bitmap in a bitmap holder due to it still potentially being
in use. We therefore add a 'drop' method to the BitmapHolder class
that sets the reference to null without recycling. This is not
ideal, but is better than recycling too early and causing crashes.