Android: Make the 'tap to skip page' regions more smartly sized.
Rather than always using 20% of the screen for the tap region
(which can be excessive and confusing on a large tablet), instead
try to use just the left/right inch.

We do this by reading the dpi of the screen. As (apparently) some
devices lie about the dpi of the screen we arrange first to always
use at least 100 pixels, and then never to use more than 1/5 of the
screen.