Bug 693527: Avoid JP2K images with themselves as their SMask.
Bug 693527 points out that we fail to spot Jpeg 2000 images that
have themselves as their own SMask and enter an infinite loop.

We extend the code by passing the forcemask parameter in exactly the
same way as the non-JPEG 2K code does.

Thanks to Jarkko Poyry for spotting this, reporting it, and suggesting
the fix.