Bug 693520: Fix rendering of non-orthogonal gradients.
Jarkko Poyry(*) points out that gradients are incorrectly rendered
when they aren't axis aligned. This review fixes it here using a
patch inspired by both his and zenikos patch. Thanks guys.

Further thanks to zeniko for spotting that it applies to the XPS
code too and providing a patch.

* Apologies for the lack of the accent - my editor/git gives problems
with them in commit messages.