Android: Refine Smart Motion to avoid tiny scrolls at end of pages
Tor observes that when advancing down a page, it's annoying to have a
tiny scroll at the end to expose the last pixel. To avoid this, he
suggests making the amount we scroll dynamic; rather than always
moving by 90% of the screen height, we allow ourselves to move between
80% and 95% of the screen height if it means that we'll exactly meet
the end of the page.

This seems to work well.