Android: Fix NullPointerException seen on Google Play
According to Google Plays automated crash detection, we get a
NullPointerException when trying to parse a null string as a
Uri.

This turns out to be caused by us trying to open a PDF attachment
from gmail. This is because MuPDF is invoked with a content:// URL
that does not have a file associated with it. Instead we can open
that URL as an InputStream.

Here we amend MuPDF to spot that case, and to open the InputStream,
suck the data into a byteArray, and then to use that to open the
file from.