Rework text extraction structures.
Rework the text extraction structures - the broad strokes are similar
but we now hold more information at each stage to enable us to perform
more detailed analysis on the structure of the page.

We now hold:

  fz_text_char's (the position, ucs value, and style of each char).

  fz_text_span's (sets of chars that share the same baseline/transform,
  with no more than an expected amount of whitespace between each char).

  fz_text_line's (sets of spans that share the same baseline (more or
  less, allowing for super/subscript, but possibly with a larger than
  expected amount of whitespace).

  fz_text_block's (sets of lines that follow one another)

After fz_text_analysis is called, we hope to have fz_text_blocks split
such that each block is a paragraph.

This new implementation has the same restrictions as the current
implementation it replaces, namely that chars are only considered for
addition onto the most recent span at the moment, but this revised form
is designed to allow more easy extension, and for this restriction to
be lifted.

Also add simple paragraph splitting based on finding the most common
'line distance' in blocks.

When we add spans together to collate them into lines, we record the
'horizontal' and 'vertical' spacing between them. (Not actually
horizontal or vertical, so much as 'in the direction of writing' and
'perpendicular to the direction of writing').

The 'horizontal' value enables us to more correctly output spaces when
converting to (say) html later.

The 'vertical' value enables us to spot subscripts and superscripts etc,
as well as small changes in the baseline due to style changes. We are
careful to base the baseline comparison on the baseline for the line,
not the baseline for the previous span, as otherwise superscripts/
subscripts on the end of the line affect what we match next.

Also, we are less tolerant of vertical shifts after a large gap. This
avoids false positives where different columns just happen to almost
line up.