Move pdf_image to fz_image.
In order to be able to output images (either in the pdfwrite device or
in the html conversion), we need to be able to get to the original
compressed data stream (or else we're going to end up recompressing
images). To do that, we need to expose all of the contents of pdf_image
into fz_image, so it makes sense to just amalgamate the two.

This has knock on effects for the creation of indexed colorspaces,
requiring some of that logic to be moved.

Also, we need to make xps use the same structures; this means pushing
PNG and TIFF support into the decoding code. Also we need to be able
to load just the headers from PNG/TIFF/JPEGs as xps doesn't include
dimension/resolution information.

Also, separate out all the fz_image stuff into fitz/res_image.c rather
than having it in res_pixmap.